Spotify - Play! It is working!
After back and forth it is working:

It was necessary though to startup a HEROKU machine for the backend authentication as Spotify /api/token does not work from within browser. Thus when access token is finally given to the client from the Heroku machine, the rest is playing easily.

Created a spotify helper to provide the necessary functions.