Blog Redesign #1
This is a first redesign of the blog to achieve a more cleaner look and feel, but keeping the personal character of it. This redesign is done with the help of Tailwind.css which helped in such a way that there where predefined examples to reference to on https://tailwindcomponents.com.

Furthermore the definite set of UI colors, font-sizes and paddings helps to keep design clean which is good. Staying in the tailwind ecosystem enforces that there aren't 20 shades of grey or a mix of em, px or rems. Nevertheless a cleanup can be done afterwards as I think several things are no longer used maybe or can be replaced.